BC-662-B | Introduced MockAmendmentFormsFunctions
