+++ src/test/java/uk/gov/justice/laa/amend/claim/views/amendments/AmendMatterTypeCodeViewTest.java
@@ -13,6 +13,7 @@ import uk.gov.justice.laa.amend.claim.controllers.amendments.AmendCaseTypeContro
 import uk.gov.justice.laa.amend.claim.forms.amendments.AmendmentForm;
 import uk.gov.justice.laa.amend.claim.forms.amendments.AmendmentForms;
 import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
+import uk.gov.justice.laa.amend.claim.resources.MockAmendmentFormsFunctions;
 import uk.gov.justice.laa.amend.claim.models.enums.AreaOfLaw;
 import uk.gov.justice.laa.amend.claim.resources.MockClaimsFunctions;
 import uk.gov.justice.laa.amend.claim.service.AvailableFeeCodesService;
@@ -42,7 +43,7 @@ class AmendMatterTypeCodeViewTest extends AmendmentsBaseTest {
     claim.setMatterType1(MATTER_TYPE_CODE_1);
     claim.setMatterType2(MATTER_TYPE_CODE_2);
 
-    var forms = createCaseTypeForm(claim);
+    var forms = MockAmendmentFormsFunctions.justCaseTypeFilled(claim);
     session.setAttribute(AMENDMENTS_KEY.formatted(claimId), forms);
 
     when(availableFeeCodesService.getAvailableFeeCodes(any())).thenReturn(Map.of(FEE_CODE, "ABC"));
@@ -70,7 +71,7 @@ class AmendMatterTypeCodeViewTest extends AmendmentsBaseTest {
     claim.setMatterType1(MATTER_TYPE_CODE_1);
     claim.setMatterType2(MATTER_TYPE_CODE_2);
 
-    var forms = createCaseTypeForm(claim);
+    var forms = MockAmendmentFormsFunctions.justCaseTypeFilled(claim);
     session.setAttribute(AMENDMENTS_KEY.formatted(claimId), forms);
 
     when(availableFeeCodesService.getAvailableFeeCodes(any())).thenReturn(Map.of(FEE_CODE, "ABC"));
@@ -87,14 +88,6 @@ class AmendMatterTypeCodeViewTest extends AmendmentsBaseTest {
     assertPageHasLabel(doc, "matter-type-two-input", "Amended matter type 2");
   }
 
-  private AmendmentForms createCaseTypeForm(ClaimDetails claim) {
-    var view = ClaimCaseViewFactory.create(claim);
-    return AmendmentForms.builder()
-        .client1(new AmendmentForm())
-        .caseType(new AmendmentForm(view.caseTypeRows()))
-        .caseDetails(new AmendmentForm())
-        .build();
-  }
 
   private void assertCommonPageContent(Document doc) {
     assertPageHasTitle(doc, "Amend claim details");
